Refactor block action types
