+++ src/main/kotlin/com/github/quiltservertools/ledger/actions/BlockChangeActionType.kt
@@ -1,20 +1,30 @@
 package com.github.quiltservertools.ledger.actions
 
 import com.github.quiltservertools.ledger.logWarn
+import com.github.quiltservertools.ledger.utility.TextColorPallet
 import com.github.quiltservertools.ledger.utility.getWorld
+import com.github.quiltservertools.ledger.utility.literal
 import net.minecraft.block.BlockState
 import net.minecraft.block.Blocks
 import net.minecraft.nbt.StringNbtReader
 import net.minecraft.network.packet.s2c.play.BlockUpdateS2CPacket
 import net.minecraft.server.MinecraftServer
 import net.minecraft.server.network.ServerPlayerEntity
+import net.minecraft.text.HoverEvent
+import net.minecraft.text.LiteralText
+import net.minecraft.text.Text
+import net.minecraft.text.TranslatableText
 import net.minecraft.util.Identifier
+import net.minecraft.util.Util
 import net.minecraft.util.registry.Registry
 
-open class BlockChangeActionType(override val identifier: String) : AbstractActionType() {
+open class BlockChangeActionType : AbstractActionType() {
+    override val identifier = "block-change"
+
     override fun rollback(server: MinecraftServer): Boolean {
         val world = server.getWorld(world)
         world?.setBlockState(pos, oldBlockState())
+        world?.getBlockEntity(pos)?.readNbt(StringNbtReader.parse(extraData))
 
         return true
     }
@@ -29,9 +39,6 @@ open class BlockChangeActionType(override val identifier: String) : AbstractActi
         val world = server.getWorld(world)
 
         world?.setBlockState(pos, newBlockState())
-        if (newBlockState().hasBlockEntity()) {
-            world?.getBlockEntity(pos)?.readNbt(StringNbtReader.parse(extraData))
-        }
 
         return true
     }
@@ -44,9 +51,43 @@ open class BlockChangeActionType(override val identifier: String) : AbstractActi
 
     override fun getTranslationType() = "block"
 
-    private fun oldBlockState() = checkForBlockState(oldObjectIdentifier, oldBlockState)
+    override fun getObjectMessage(): Text {
+        val text = LiteralText("")
+        text.append(
+            TranslatableText(
+            Util.createTranslationKey(
+                this.getTranslationType(),
+                oldObjectIdentifier
+            )
+        ).setStyle(TextColorPallet.secondaryVariant).styled {
+            it.withHoverEvent(
+                HoverEvent(
+                    HoverEvent.Action.SHOW_TEXT,
+                    oldObjectIdentifier.toString().literal()
+                )
+            )
+        })
+        text.append("→".literal())
+        text.append(
+            TranslatableText(
+            Util.createTranslationKey(
+                this.getTranslationType(),
+                objectIdentifier
+            )
+        ).setStyle(TextColorPallet.secondaryVariant).styled {
+            it.withHoverEvent(
+                HoverEvent(
+                    HoverEvent.Action.SHOW_TEXT,
+                    objectIdentifier.toString().literal()
+                )
+            )
+        })
+        return text
+    }
 
-    private fun newBlockState() = checkForBlockState(objectIdentifier, blockState)
+    fun oldBlockState() = checkForBlockState(oldObjectIdentifier, oldBlockState)
+
+    fun newBlockState() = checkForBlockState(objectIdentifier, blockState)
 
     private fun checkForBlockState(identifier: Identifier, checkState: BlockState?): BlockState {
         val block = Registry.BLOCK.getOrEmpty(identifier)
