+++ src/main/resources/data/ledger/lang/en_us.json
@@ -47,10 +47,10 @@
 
     "text.ledger.action.block-place": "placed",
     "text.ledger.action.block-break": "broke",
+    "text.ledger.action.block-change": "changed",
     "text.ledger.action.item-insert": "added",
     "text.ledger.action.item-remove": "removed",
     "text.ledger.action.entity-kill": "killed",
-    "text.ledger.action.block-interact": "interacted with",
 
     "text.ledger.network.protocols_mismatched": "Protocol versions mismatched: client mod is Ledger protocol version %d and the client version is %d",
     "text.ledger.network.no_mod_info": "Unable to determine mod information for your client mod",
